backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
pymysql
requests
flasgger
python-dotenv
